requirements.txt
django~=4.0.1
cfgv
virtualenv
liccheck
djangorestframework~=3.13.1